% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_cleaned_excel_with_elog.R
\name{read_in_cleaned_excel_with_elog}
\alias{read_in_cleaned_excel_with_elog}
\title{Title}
\usage{
read_in_cleaned_excel_with_elog(file_path)
}
\arguments{
\item{file_path}{}
}
\description{
Title
}
\author{
Muhammad Abdullah
}
