% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_cleaned_excel_without_elog.R
\name{read_in_cleaned_excel_without_elog}
\alias{read_in_cleaned_excel_without_elog}
\title{Title}
\usage{
read_in_cleaned_excel_without_elog(file_path)
}
\description{
Title
}
\author{
Muhammad Abdullah
}
